# swagger_API.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Hotel Booking API",
      "description": "API endpoints for authentication, user management, payments, and feedback in the hotel booking system.",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "paths": {
...
      "/api/feedback": {
        "post": {
          "tags": [
            "Feedback"
          ],
          "summary": "Add feedback for a hotel",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddFeedbackRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Feedback added successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Feedback"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid feedback data or hotel not found"
            },
            "403": {
              "description": "Only travelers can add feedback"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      },
      "/api/feedback/{hotel_id}": {
        "get": {
          "tags": [
            "Feedback"
          ],
          "summary": "Get all feedback for a specific hotel",
          "parameters": [
            {
              "name": "hotel_id",
              "in": "path",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Returns feedback list for the hotel",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Feedback"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Hotel not found"
            }
          }
        }
      },
      "/api/feedback/{id}": {
        "delete": {
          "tags": [
            "Feedback"
          ],
          "summary": "Delete feedback by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Feedback deleted successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DeleteFeedbackResponse"
                  }
                }
              }
            },
            "403": {
              "description": "Only hotel managers can delete feedback"
            },
            "404": {
              "description": "Feedback not found"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Feedback": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "hotel_id": {
              "type": "integer"
            },
            "traveler_id": {
              "type": "string",
              "format": "uuid"
            },
            "comments": {
              "type": "string"
            },
            "rating": {
              "maximum": 5,
              "minimum": 1,
              "type": "integer"
            }
          }
        },
        "AddFeedbackRequest": {
          "required": [
            "hotel_id",
            "rating",
            "traveler_id"
          ],
          "type": "object",
          "properties": {
            "hotel_id": {
              "type": "integer"
            },
            "traveler_id": {
              "type": "string",
              "format": "uuid"
            },
            "comments": {
              "type": "string"
            },
            "rating": {
              "maximum": 5,
              "minimum": 1,
              "type": "integer"
            }
          }
        },
        "DeleteFeedbackResponse": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }